writing tests that mock
In this commit with `patch('django.db.utils.ConnectionHandler.__getitem__') as gi:`

In the django base code this 'get_item' is the function being called that actual connects to the db. We are mocking it here to True. Call count is attached to all these methods and allows us to keep track of how many times they were called.